requirements.txt
openai>=1.0.0,<2
chromadb>=0.5.0
pydantic>=2.0.0
pdfplumber>=0.10.0
playwright>=1.45.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
numpy>=1.26.0
